+++ src/game/boards/random/index.js
@@ -1,0 +1,11 @@
+import board1 from "./board1.json"
+import board2 from "./board2.json"
+import board3 from "./board3.json"
+import board4 from "./board4.json"
+
+export default [
+    {name: "Board 1", ...board1},
+    {name: "Board 2", ...board2},
+    {name: "Board 3", ...board3},
+    {name: "Board 4", ...board4}
+]
