+++ src/game/boards/relaxed/index.js
@@ -1,0 +1,5 @@
+import board1 from './board1.json'
+
+export default [
+    {name: "Board 1", ...board1}
+]
